requirements.txt
flask
cnocr
pillow
numpy
pandas
opencv-python